Surface pagination-aware errors
Add a user repository pagination error surface so invalid cursor and
unsupported direction failures can travel through the port layer without
being collapsed into generic persistence query failures.

Classify unsupported cursor directions in the pagination crate, map both
repository-originated and inbound cursor failures to HTTP 400 details, and
cover the externally visible users-list behaviour with rstest-bdd.
