+++ backend/tests/users_list_pagination_bdd.rs
@@ -58,6 +58,19 @@ fn the_client_requests_the_users_list_with_an_invalid_cursor(world: &mut World) 
     run_authenticated_request(world, "/api/v1/users?cursor=not-a-cursor");
 }
 
+#[when("the client requests the users list with an unsupported cursor direction")]
+fn the_client_requests_the_users_list_with_an_unsupported_cursor_direction(world: &mut World) {
+    run_authenticated_request(
+        world,
+        concat!(
+            "/api/v1/users?cursor=",
+            "eyJrZXkiOnsiY3JlYXRlZF9hdCI6IjIwMjYtMDEtMDFUMDA6MDA6MDBaIiwiaWQi",
+            "OiIxMTExMTExMS0xMTExLTExMTEtMTExMS0xMTExMTExMTExMTEifSwiZGlyIjoi",
+            "U2lkZXdheXMifQ"
+        ),
+    );
+}
+
 #[when("the client requests the users list without a session")]
 fn the_client_requests_the_users_list_without_a_session(world: &mut World) {
     run_unauthenticated_request(world);
@@ -108,6 +121,11 @@ fn the_users_response_is_bad_request_with_invalid_cursor_details(world: &mut Wor
     assert_error(world, 400, "invalid_cursor");
 }
 
+#[then("the users response is bad request with unsupported_direction details")]
+fn the_users_response_is_bad_request_with_unsupported_direction_details(world: &mut World) {
+    assert_error(world, 400, "unsupported_direction");
+}
+
 #[then("the users response is unauthorised")]
 fn the_users_response_is_unauthorised(world: &mut World) {
     assert_status(world, 401);
@@ -155,6 +173,14 @@ fn oversized_users_page_limit_is_rejected(world: World) {
     name = "Invalid users cursor is rejected"
 )]
 fn invalid_users_cursor_is_rejected(world: World) {
+    drop(world);
+}
+
+#[scenario(
+    path = "tests/features/users_list_pagination.feature",
+    name = "Unsupported users cursor direction is rejected"
+)]
+fn unsupported_users_cursor_direction_is_rejected(world: World) {
     drop(world);
 }
